backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
playwright
pdfplumber
python-dotenv
pydantic
aiofiles
fpdf2
